+++ packages/ketcher-core/src/application/render/coordinateTransformation.ts
@@ -1,5 +1,5 @@
 import { Vec2 } from 'domain/entities';
-import { Render } from './raphaelRender';
+import type { Render } from './raphaelRender';
 import { Scale } from 'domain/helpers';
 
 const canvasToView = (point: Vec2, render: Render) => {
