+++ packages/ketcher-core/src/application/render/restruct/generalEnumTypes.ts
@@ -45,3 +45,11 @@ export enum StereLabelStyleType {
   On = 'On',
   Off = 'Off',
 }
+
+export enum ShowHydrogenLabels {
+  Off = 'off',
+  Hetero = 'Hetero',
+  Terminal = 'Terminal',
+  TerminalAndHetero = 'Terminal and Hetero',
+  On = 'all',
+}
