+++ packages/ketcher-core/src/application/render/restruct/rebond.ts
@@ -28,7 +28,7 @@ import { getColorFromStereoLabel } from './reatom';
 
 import ReObject from './reobject';
 import ReStruct from './restruct';
-import { Render } from '../raphaelRender';
+import type { Render } from '../raphaelRender';
 import { Scale } from 'domain/helpers';
 import draw from '../draw';
 import util from '../util';
