+++ packages/ketcher-core/src/application/render/restruct/reobject.ts
@@ -16,7 +16,7 @@
 
 import { Box2Abs } from 'domain/entities';
 import ReStruct from './restruct';
-import { Render } from '../raphaelRender';
+import type { Render } from '../raphaelRender';
 import { Scale } from 'domain/helpers';
 import Visel from './visel';
 import { IMAGE_KEY } from 'domain/constants';
