+++ packages/ketcher-core/src/application/render/restruct/resimpleObject.ts
@@ -19,7 +19,7 @@ import { Box2Abs, SimpleObjectMode, Vec2 } from 'domain/entities';
 import { LayerMap } from './generalEnumTypes';
 import ReObject from './reobject';
 import ReStruct from './restruct';
-import { Render } from '../raphaelRender';
+import type { Render } from '../raphaelRender';
 import { Scale } from 'domain/helpers';
 import draw from '../draw';
 import util from '../util';
