+++ packages/ketcher-core/src/application/render/scrollbar/scrollbar-container.ts
@@ -1,4 +1,4 @@
-import { Render } from '../raphaelRender';
+import type { Render } from '../raphaelRender';
 import { ScrollOffset } from './scroll-offset';
 import { VerticalScrollbar } from './scrollbar-vertical';
 import { HorizontalScrollbar } from './scrollbar-horizontal';
